requirements.txt
requests
phonenumbers
beautifulsoup4
lxml
rich
pyfiglet
geopy
folium
streamlit
python-dotenv
